+++ src/DeauthD/deauth_detect.h
@@ -8,7 +8,12 @@ extern unsigned long total_attack_counter; // Total attacks counter
 extern bool ATTACK;                        // is attack currently happening
 extern int packets_count;                 // Last Deauth packets counts
 
+#ifdef PLATFORM_8266
 void sniffer(uint8_t *buf, uint16_t len);
+#endif
+#ifdef PLATFORM_ESP32
+void sniffer(void* buf, wifi_promiscuous_pkt_type_t type); 
+#endif 
 void attack_started();
 void attack_stopped();