+++ ChangeLog.md
@@ -1,3 +1,8 @@
+-=# v0.0.5 (20260209) #=- 
+ADD: 
+ * Multiplatform in .ini (Deauth on ESP32 (in progress)/ESP8266 (working)) 
+ * Start adding support for MiniTV 1.44" 
+
 -=# v0.0.4 (20260208) #=- 
 ADD: 
  * SERIAL_DEBUG on/off